+++ dbrequests/mysql/tests/conftest.py
@@ -101,6 +101,18 @@ def set_up_cats(db):
         """)
 
 
+def set_up_diffs(db):
+    db.bulk_query("""
+    CREATE TABLE IF NOT EXISTS `diffs` (
+        `id` bigint(20) NOT NULL AUTO_INCREMENT,
+        `value` varchar(100) NOT NULL,
+        `updated` timestamp NOT NULL DEFAULT current_timestamp(),
+        PRIMARY KEY (`id`)
+    ) ENGINE=InnoDB AUTO_INCREMENT=2
+    """)
+    db.bulk_query('TRUNCATE TABLE diffs')
+
+
 def set_up_membership(db):
     db.bulk_query("""
     CREATE TABLE IF NOT EXISTS `membership`
@@ -120,7 +132,7 @@ def set_up_membership(db):
     """)
 
 
-@pytest.fixture(scope="module", autouse=True)
+@pytest.fixture(scope="package", autouse=True)
 def container_controller(request):
     """Startup database fixture."""
     container = run_docker_container()